+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.10.1.0-beta] - 2024-XX-XX
+## [0.11.0.0] - 2024-XX-XX
 New release of the HyperDbg Debugger.
 
 ### Added
@@ -14,6 +14,8 @@ New release of the HyperDbg Debugger.
 - The 'struct' command now supports a path as output ([link](https://docs.hyperdbg.org/commands/debugging-commands/struct))
 - Export SDK API closing connection to the remote debuggee
 - Automated tests for the main command parser
+- Export SDK APIs for stepping and tracing instructions
+- Export SDK APIs for tracking execution
 
 ### Changed
 - HyperDbg command-line comment sign is changed from '#' to C-like comments ('//' and '/**/')